pyproject.toml
[project]
name = "streamline-vocals"
version = "0.1.0"
description = "Vocal Transformation UI — ACE-Step remix pipeline with pitch shifting and RVC"
requires-python = ">=3.11"
dependencies = [
    # Torch stack pinned to PyTorch 2.10 with CUDA 13.0 wheels.
    "torch==2.10.0+cu130",
    "torchaudio==2.10.0+cu130",
    "torchvision==0.25.0+cu130",
    # Server
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "loguru>=0.7.0",
    # Audio / transcription
    "openai-whisper>=20240930",
    "pyworld>=0.3.4",
    "ffmpeg-python>=0.2.0",
    # Pinned to applio's required versions to avoid silent incompatibilities
    "numpy==2.4.4",
    "scipy==1.17.1",
    "soundfile==0.13.1",
    # RVC / Applio inference dependencies
    "faiss-cpu==1.13.2",
    "librosa==0.11.0",
    "noisereduce",
    "pedalboard",
    "stftpitchshift",
    "soxr",
    "numba==0.65.0",
    "torchcrepe",
    "torchfcpe",
    "einops",
    "transformers==5.4.0",
    "omegaconf>=2.0.6",
    "wget",
    "requests",
    "tqdm",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["streamline_svc"]

[dependency-groups]
dev = []

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu130" }]
torchaudio = [{ index = "pytorch-cu130" }]
torchvision = [{ index = "pytorch-cu130" }]
